refactor: implement HttpClient as a singleton
