+++ app/src/main/java/com/coderGtm/yantra/commands/weather/Command.kt
@@ -1,19 +1,9 @@
 package com.coderGtm.yantra.commands.weather
 
-import android.util.Log
-import com.coderGtm.yantra.BuildConfig
 import com.coderGtm.yantra.R
 import com.coderGtm.yantra.blueprints.BaseCommand
 import com.coderGtm.yantra.models.CommandMetadata
 import com.coderGtm.yantra.terminal.Terminal
-import io.ktor.client.HttpClient
-import io.ktor.client.engine.android.Android
-import io.ktor.client.plugins.contentnegotiation.ContentNegotiation
-import io.ktor.client.plugins.logging.LogLevel
-import io.ktor.client.plugins.logging.Logging
-import io.ktor.serialization.kotlinx.json.json
-import io.ktor.util.logging.Logger
-import kotlinx.serialization.json.Json
 
 class Command(terminal: Terminal) : BaseCommand(terminal) {
     override val metadata = CommandMetadata(
@@ -22,28 +12,6 @@ class Command(terminal: Terminal) : BaseCommand(terminal) {
         description = terminal.activity.getString(R.string.cmd_weather_help)
     )
 
-    // todo: ideally this would be a singleton that is re-used across the app
-    private val httpClient = HttpClient(Android) {
-        expectSuccess = true
-
-        install(ContentNegotiation) {
-            json(Json {
-                ignoreUnknownKeys = true
-                coerceInputValues = true
-            })
-        }
-        if (BuildConfig.DEBUG) {
-            install(Logging) {
-                logger = object : io.ktor.client.plugins.logging.Logger {
-                    override fun log(message: String) {
-                        Log.d("HTTP call", message)
-                    }
-                }
-                level = LogLevel.ALL
-            }
-        }
-
-    }
 
     override fun execute(command: String) {
         when (val parseResult = parseWeatherCommand(command, this.terminal.activity)) {
@@ -55,7 +23,7 @@ class Command(terminal: Terminal) : BaseCommand(terminal) {
             )
 
             is ParseResult.ListCommand -> showAvailableFields(this)
-            is ParseResult.Success -> fetchWeatherData(parseResult.args, this, httpClient)
+            is ParseResult.Success -> fetchWeatherData(parseResult.args, this)
         }
     }
 }
